Split off separate files in download lib
