+++ src/UnifiedUpdatePlatform.Services.WindowsUpdate.Downloads/HttpDownloader.cs
@@ -35,71 +35,6 @@ using UnifiedUpdatePlatform.Services.WindowsUpdate.Models.FE3.JSON.ESRP;
 
 namespace UnifiedUpdatePlatform.Services.WindowsUpdate.Downloads
 {
-    public class GeneralDownloadProgress
-    {
-        public long EstimatedTotalBytes;
-        public long DownloadedTotalBytes;
-        public int NumFilesDownloadedSuccessfully;
-        public int NumFilesDownloadedUnsuccessfully;
-        public int NumFiles;
-
-        public FileDownloadStatus[] DownloadedStatus;
-    }
-
-    public enum FileStatus
-    {
-        Downloading,
-        Verifying,
-        Completed,
-        Expired,
-        Failed
-    }
-
-    public class FileDownloadStatus
-    {
-        public FileStatus FileStatus;
-        public long DownloadedBytes;
-        public long HashedBytes;
-        public UUPFile File;
-        public FileDownloadStatus(UUPFile file)
-        {
-            File = file;
-        }
-    }
-
-    public class UUPFile
-    {
-        public FileExchangeV3FileDownloadInformation WUFile
-        {
-            get; set;
-        }
-        public string FileName
-        {
-            get; set;
-        }
-        public long FileSize
-        {
-            get; set;
-        }
-        public string Hash
-        {
-            get; set;
-        }
-        public string HashAlgorithm
-        {
-            get; set;
-        }
-
-        public UUPFile(FileExchangeV3FileDownloadInformation WUFile, string FileName, long FileSize, string Hash, string HashAlgorithm)
-        {
-            this.WUFile = WUFile;
-            this.FileName = FileName;
-            this.FileSize = FileSize;
-            this.Hash = Hash;
-            this.HashAlgorithm = HashAlgorithm;
-        }
-    }
-
     public class HttpDownloader : IDisposable
     {
         private const long CHUNK_SIZE = 8_388_608 + 65_536; //Slice 8MB+64KB
