+++ src/UnifiedUpdatePlatform.Services.WindowsUpdate.Downloads/UUPFile.cs
@@ -1,0 +1,57 @@
+﻿/*
+ * Copyright (c) Gustave Monce and Contributors
+ * Copyright (c) ADeltaX and Contributors
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+namespace UnifiedUpdatePlatform.Services.WindowsUpdate.Downloads
+{
+    public class UUPFile
+    {
+        public FileExchangeV3FileDownloadInformation WUFile
+        {
+            get; set;
+        }
+        public string FileName
+        {
+            get; set;
+        }
+        public long FileSize
+        {
+            get; set;
+        }
+        public string Hash
+        {
+            get; set;
+        }
+        public string HashAlgorithm
+        {
+            get; set;
+        }
+
+        public UUPFile(FileExchangeV3FileDownloadInformation WUFile, string FileName, long FileSize, string Hash, string HashAlgorithm)
+        {
+            this.WUFile = WUFile;
+            this.FileName = FileName;
+            this.FileSize = FileSize;
+            this.Hash = Hash;
+            this.HashAlgorithm = HashAlgorithm;
+        }
+    }
+}
